// apps/web/src/lib/generated/models/SourceAssetPageRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { SourceAssetRead } from './SourceAssetRead';
export type SourceAssetPageRead = {
  items: Array<SourceAssetRead>;
  limit: number;
  offset: number;
  total: number;
};